gourd_wrapper::measurement_unix

Trait GetRUsage

pub trait GetRUsage {
    // Required method
    fn wait_for_rusage(&self) -> Result<(Option<RUsage>, i32), Error>;
}

A trait for getting resource usage statistics for a process.

Required Methods

fn wait_for_rusage(&self) -> Result<(Option<RUsage>, i32), Error>

Waits for the process to exit and returns its resource usage statistics. Works only on linux with wait4 syscall available.